+++ models/project.js
@@ -68,6 +68,9 @@ var ProjectSchema = new Schema({
     type: String,
     select: false
   },
+  attributes: {
+    type: Object,
+  },
   // apiKey: { //You do want to block anonymous traffic. https://cloud.google.com/endpoints/docs/openapi/when-why-api-key#:~:text=API%20keys%20identify%20an%20application's,patterns%20in%20your%20API's%20traffic
   //   type: String,
   //   select: false,