+++ pubmodules/emailNotification/requestNotification.js
@@ -24,7 +24,7 @@ var configGlobal = require('../../config/global');
 
 var widgetConfig = require('../../config/widget');
 var widgetTestLocation = process.env.WIDGET_TEST_LOCATION || widgetConfig.testLocation;
-let configSecret = process.env.GLOBAL_SECRET || config.secret;
+let configSecret = process.env.GLOBAL_SECRET_OR_PRIVATE_KEY || process.env.GLOBAL_SECRET || config.secret;
 
 let apiUrl = process.env.API_URL || configGlobal.apiUrl;
 winston.debug('********* RequestNotification apiUrl: ' + apiUrl);
@@ -290,7 +290,8 @@ sendToUserEmailChannelEmail(projectid, message) {
         issuer:  'https://tiledesk.com',
         subject:  'userexternal',
         audience:  'https://tiledesk.com',
-        jwtid: uuidv4()        
+        jwtid: uuidv4(),
+        algorithm: process.env.GLOBAL_SECRET_ALGORITHM        
       };
 
 
@@ -301,7 +302,7 @@ sendToUserEmailChannelEmail(projectid, message) {
       winston.debug("userEmail  ",userEmail);
 
 
-      var token = jwt.sign(userEmail, configSecret, signOptions);
+      var token = jwt.sign(userEmail, configSecret, signOptions); //priv_jwt pp_jwt
       winston.debug("token  "+token);
 
       var sourcePage = widgetTestLocation + "?tiledesk_projectid=" 
@@ -769,14 +770,15 @@ sendUserEmail(projectid, message) {
                     issuer:  'https://tiledesk.com',
                     subject:  'guest',
                     audience:  'https://tiledesk.com',
-                    jwtid: uuidv4()        
+                    jwtid: uuidv4(),
+                    algorithm: process.env.GLOBAL_SECRET_ALGORITHM 
                   };
 
                   let userAnonym = {_id: recipient, firstname: lead.fullname, lastname: lead.fullname, email: lead.email, attributes: lead.attributes};
                   winston.debug("userAnonym  ",userAnonym);
 
         
-                  var token = jwt.sign(userAnonym, configSecret, signOptions);
+                  var token = jwt.sign(userAnonym, configSecret, signOptions); //priv_jwt pp_jwt
                   winston.debug("token  "+token);
 
                   var sourcePage = widgetTestLocation + "?tiledesk_projectid="